requirements.txt
mysqlclient
mysql-connector